Handle ephemeral pages
Now that there are no more visible tabs, there will ever only be a single
application open at a time. There is no button or link displayed for the app
page. Towards that end, I added the concept of "ephemeral pages" which are
indicated by prepending a dash to the page label. Ephemeral pages are
displayed without a button, and are removed when any other page, including
a new ephemeral page, is displayed. (The latter allows for the possibility of
an app page containing "similar apps" and moving from app to app without ever
moving back to Home of Find Apps.

The hierarchy indicator now displays the title of the app adjacent to the page
from which it came... unless it came from another ephemeral page, in which
case it appears to have come from the most recent non-ephemeral page. (There
is not yet any case of ephemeral-to-ephemeral movement.)

Derrell